Allow array index to be uint256
This goes back to commit 444870548f1008b9c6b42723ae8fdcb367acbf97.
Before that, the type of the index was inferred. After that, it changed
to be the default numeric type in the language, `num` (now `int128`).

Array index types should generally be uint256.

+++ vyper/parser/parser_utils.py
@@ -334,7 +334,7 @@ def add_variable_offset(parent, key, pos):
 
         subtype = typ.subtype
         sub = [
-            'uclamplt', base_type_conversion(key, key.typ, BaseType('int128'), pos=pos), typ.count
+            'uclamplt', base_type_conversion(key, key.typ, BaseType('uint256'), pos=pos), typ.count
         ]
 
         if location == 'storage':
